fix: instance code display (#868)
## Context <!-- ie. explanations, background, documentation -->

- instance codes above menu was wrongly displayed its all uppercase
- changed code font to grant lI and O0 differentiation
- added copy icon and feedback on clicking on code copy button

## Checklist

- [x] Tested manually <!-- you can strikethrough this option in case you
haven't tested manually -->
- [ ] Passed automatic tests <!-- you can strikethrough this option in
case you haven't run the automatic tests for some reason -->
- [x] GitHub issue linked <!-- Use the "Development" field of the Issue,
or add a link if it's outside this Repo -->
- [x] Changelist updated
- [x] Backward and forward compatible with BE <!-- If not, please
describe in detail and include other PR links -->
- [x] Doesn't need update in the database or BE <!-- If it does, please
describe how to deploy it without downtime -->
- [ ] Must be deployed ASAP (HOTFIX)
