+++ public/font/inconsolata/inconsolata.css
@@ -1,0 +1,5 @@
+@font-face {
+    font-family: 'inconsolata';
+    src: url('/font/inconsolata/inconsolata.woff') format('woff');
+    font-style: normal;
+}
