+++ src/components/Buttons/CodeButton/index.tsx
@@ -1,0 +1,3 @@
+import CodeButton from './CodeButton';
+
+export { CodeButton as default };
